feat: use socket io for scan result
close #77 close #78

+++ frontend/src/hooks/useScanResults.tsx
@@ -4,8 +4,14 @@ import { fileStatus } from "types/file";
 import { Event } from "types/events";
 import { match } from "ts-pattern";
 import { useFileUploader } from "components/providers/fileUploader";
+import debug from "debug";
+import { io } from "socket.io-client";
+import { useEffect, useRef } from "react";
+import { useLocalStorage } from "./useLocalStorage";
+const log = debug("useScanResults");
 
 export const useScanResults = () => {
+  log("useScanResults");
   const { emit, events$, clean } = useEvents({
     key: "scanResults",
     eventKind: (s: any) => {
@@ -18,18 +24,26 @@ export const useScanResults = () => {
 
   const { removeFile } = useFileUploader();
 
-  useSWRSubscription("scanResult", (key, { next }) => {
-    const eventSource = new WebSocket(
-      `${import.meta.env.VITE_BACKEND_WS_URL}/scan-result`
-    );
+  const socketRef = useRef(
+    io(`${import.meta.env.VITE_BACKEND_WS_URL}/scan-result`, {
+      autoConnect: false,
+    })
+  );
 
-    eventSource.addEventListener("message", (event) => {
-      const file = JSON.parse(event.data);
-      removeFile(file.id);
-      emit(file, file.result);
-    });
+  const [accessToken] = useLocalStorage("accessToken");
 
-    return () => eventSource.close();
+  useEffect(() => {
+    if (accessToken) {
+      socketRef.current.auth = { token: accessToken };
+      socketRef.current.connect();
+      console.log("connecting ");
+      socketRef.current.on("event", (event) => {
+        log("message", event);
+        const file = JSON.parse(event);
+        removeFile(file.id);
+        emit(file, file.result);
+      });
+    }
   });
 
   return {
